Refactor paramsets/paramview into parameters (#550)
* Clean up paramsets/paramview namespace by putting it under parameters
