+++ src/pyhf/parameters/__init__.py
@@ -1,0 +1,17 @@
+from .paramsets import (
+    paramset,
+    unconstrained,
+    constrained_by_normal,
+    constrained_by_poisson,
+)
+from .utils import reduce_paramsets_requirements
+from .paramview import ParamViewer
+
+__all__ = [
+    'paramset',
+    'unconstrained',
+    'constrained_by_normal',
+    'constrained_by_poisson',
+    'reduce_paramsets_requirements',
+    'ParamViewer',
+]
