+++ tests/test_constraints.py
@@ -1,6 +1,6 @@
 import pytest
 import pyhf
-from pyhf.paramsets import constrained_by_poisson, constrained_by_normal
+from pyhf.parameters import constrained_by_poisson, constrained_by_normal
 from pyhf.constraints import gaussian_constraint_combined, poisson_constraint_combined
 from pyhf import default_backend
 import numpy as np
